FEAT: change date formatter function to use short month
FIX: format date in prisoners list

FEAT: Disaply readiness section

FIX: add 'days' to profileHeader

+++ server/utils/utils.ts
@@ -27,7 +27,7 @@ export const initialiseName = (fullName?: string): string | null => {
 
 export const formatDate = (dateString: string): string => {
   const date = new Date(dateString)
-  const options: Intl.DateTimeFormatOptions = { day: 'numeric', month: 'long', year: 'numeric' }
+  const options: Intl.DateTimeFormatOptions = { day: 'numeric', month: 'short', year: 'numeric' }
   return date.toLocaleDateString('en-GB', options)
 }
